Centralizes OpenAI client creation
Refactors the creation of OpenAI and AzureOpenAI clients into a single `createOpenAIClient` function.

This change promotes code reuse and simplifies configuration across different components.
